# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "FieldFuze Backend API with DynamoDB and Telnyx Integration\n\n## 🔥 AUTHENTICATION FLOW:\n### Step 1: Register Customer\n**POST /auth/register** - Create customer account (no token generated)\n`{\"email\": \"user@example.com\", \"username\": \"john\", \"password\": \"pass123\", \"first_name\": \"John\", \"last_name\": \"Doe\"}`\n\n## 🚀 QUICK AUTHENTICATION:\n### Using the Authorize Button (Recommended)\n1. Click the **🔓 Authorize** button (top right of any API section)\n2. In the authorization dialog, use the **Login** form:\n- Enter your **Username** (email)\n- Enter your **Password**\n- Click **Login** button\n3. Your Bearer token will be **automatically applied** to all API calls!\n4. All protected endpoints will now work without manual token entry\n\n### Manual Token Entry (Alternative)\nIf you prefer manual setup: Get token from `/auth/login`, then paste `Bearer YOUR_TOKEN` in the Authorization field",
        "title": "FieldFuze Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1/auth",
    "paths": {
...
        "/organization": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new organization with specified details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organization Management"
                ],
                "summary": "Create a new organization",
                "parameters": [
                    {
                        "description": "Create organization request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Organization"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Organization created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid organization data",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict - Organization already exists",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Organization creation failed",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/user/organization": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of all organizations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organization Management"
                ],
                "summary": "Get all organizations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of organizations per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by organization status (active, inactive, archived)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organizations retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to retrieve organizations",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.APIError": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "More context about the error",
                    "type": "string"
                },
                "field": {
                    "description": "For validation errors (which field failed)",
                    "type": "string"
                },
                "type": {
                    "description": "e.g., \"ValidationError\", \"DatabaseError\"",
                    "type": "string"
                }
            }
        },
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code (200, 400, 500, etc.)",
                    "type": "integer"
                },
                "data": {
                    "description": "Any response data (can be map, struct, list, etc.)"
                },
                "error": {
                    "description": "Detailed error info (nil if success)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.APIError"
                        }
                    ]
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string"
                }
            }
        },
        "models.Organization": {
            "type": "object",
            "required": [
                "created_by",
                "name",
                "status"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 200
                },
                "city": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "country": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "industry": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "phone": {
                    "type": "string"
                },
                "postal_code": {
                    "description": "Business details",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "state": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                },
                "status": {
                    "enum": [
                        "active",
                        "inactive",
                        "suspended"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.OrganizationStatus"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "description": "Contact information",
                    "type": "string"
                }
            }
        },
        "models.OrganizationStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive",
                "suspended"
            ],
            "x-enum-varnames": [
                "OrganizationStatusActive",
                "OrganizationStatusInactive",
                "OrganizationStatusSuspended"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}